refactor(scripts): extract shared SDH3 history backfill helper
CI feedback from PR #1454 (SonarCloud: 25.8% duplication on new code,
Gemini: redundant filter + sort, CodeRabbit: doc inaccuracy).

The IRH3 backfill script was a near-clone of backfill-hah3-history.ts —
same source, same parse helper, same timezone, varying only by kennel
name + code. Extract the shared logic into scripts/lib/sdh3-history-backfill.ts
so both kennels (and the remaining 8 SDH3-fed kennels) can be backfilled
through a single thin wrapper.

Changes:
  - New scripts/lib/sdh3-history-backfill.ts exposing
    `backfillSdh3HistoryKennel({ kennelCode, kennelDisplayName, label })`.
  - Refactor scripts/backfill-hah3-history.ts: 87 → 41 lines, thin wrapper.
  - Refactor scripts/backfill-irh3-history.ts: 84 → 44 lines, thin wrapper.

Review-comment fixes:
  - Drop the defense-in-depth `kennelTags[0] === KENNEL_CODE` filter — the
    `parseHistoryEvents` kennelNameMap already filters at the source.
  - Drop the per-script `sort` — `reportAndApplyBackfill` sorts by date
    internally before sampling.
  - Use `console.warn` for the per-fetch summary so it lands in the
    warning stream where backfill-script logs are expected to appear.
  - Fix IRH3 doc header: "since 2007" → "back to Dec 2006", "462 total"
    → "346 total" (matches live parse output).

HAH3 behavior is unchanged (re-run produces same output as before). IRH3
behavior is unchanged (re-verified: 346 events, 2006-12-08 → 2026-04-03,
same sample rows).
